updating scripts
