+++ client/nodejs/src/register_shard_eth.js
@@ -30,9 +30,10 @@ const api = await ApiPromise.create({
     }
 });
 
-const time1 = "0x78af33d076b81fddce1c051a72bb1a23fd32519a2ede7ba7a54b2c76d110c54d";
-const time2 = "0xcee262950a61e921ac72217fd5578c122bfc91ba5c0580dbfbe42148cf35be2b";
-const time3 = "0xa01b6ceec7fb1d32bace8ffcac21ffe6839d3a2ebe26d86923be9dd94c0c9a02";
+const time1 = [76, 181, 171, 246, 173, 121, 251, 245, 171, 188, 202, 252, 194, 105, 216, 92, 210, 101, 30, 212, 184, 133, 181, 134, 159, 36, 26, 237, 240, 165, 186, 41];
+const time2 = [116, 34, 185, 136, 117, 152, 6, 142, 50, 196, 68, 138, 148, 154, 219, 41, 13, 15, 78, 53, 185, 224, 27, 14, 229, 241, 161, 230, 0, 254, 38, 116];
+const time3 = [243, 129, 98, 110, 65, 231, 2, 126, 164, 49, 191, 227, 0, 158, 148, 189, 210, 90, 116, 107, 238, 196, 104, 148, 141, 108, 60, 124, 93, 201, 165, 75];
+const collector_pubkey = { "sr25519": "0x78af33d076b81fddce1c051a72bb1a23fd32519a2ede7ba7a54b2c76d110c54d" };
 const phrase = "//Alice";
 const kv = new Keyring({type: 'sr25519'});
 const pair = kv.addFromUri(phrase);
@@ -43,32 +44,13 @@ const validator_phrase = "//Alice//stash";
 const validator_kv = new Keyring({ type: 'sr25519' });
 const validator_pair = validator_kv.addFromUri(validator_phrase);
 
-const register_chronicle = await api.tx.tesseractSigStorage.registerChronicle(time1).signAndSend(validator_pair, ({ status, events, dispatchError }) => {
-    console.log(`Current status is ${status}`);
-});
 
-setTimeout(async () => {
-    const register_chronicle_2 = await api.tx.tesseractSigStorage.registerChronicle(time2).signAndSend(validator_pair, ({ status, events, dispatchError }) => {
-        console.log(`Current status is ${status}`);
+const register_shard = await api.tx.sudo
+    .sudo(
+        api.tx.shards.registerShard(0, [time1, time2, time3], collector_pubkey)
+    )
+    .signAndSend(pair, (result) => {
+        console.log('Result of shard creation: ', result)
     });
-}, 15000)
-
-
-setTimeout(async () => {
-    const register_chronicle_3 = await api.tx.tesseractSigStorage.registerChronicle(time3).signAndSend(validator_pair, ({ status, events, dispatchError }) => {
-        console.log(`Current status is ${status}`);
-    });
-}, 30000)
-
-
-setTimeout(async () => {
-    const register_shard = await api.tx.sudo
-        .sudo(
-            api.tx.tesseractSigStorage.registerShard([time1, time2, time3], 0, 0)
-        )
-        .signAndSend(pair, (result) => {
-            console.log('Result of shard creation: ', result)
-        });
-}, 45000)
 
 await chan.get().then(value => console.log(value), error => console.error(error));
